Fix rhythm section note highlighting
